Use the custom 0.3.3 version of the geofences plugin
This is done for three motives, first a couple of custom
modification were made in the plugin in order to
capture the notification data, second the receiveTransition function
has been renamed to onReceiveTransition in the >=0.4 version, and third
the promises plugin dependency was not fixed and with the switch
of cordova 5 to gradle many plugins has been renamed ie.
com.vladstirbu.cordova.promise to es6-promise-plugin.Promise causing
errors in the plugin installation.